+++ src/lib/pages/validator-details/components/validator-top/ValidatorStats.tsx
@@ -6,7 +6,7 @@ import {
   useValidatorStakingProvisions,
 } from "lib/services/validatorService";
 import type { Option, ValidatorAddr } from "lib/types";
-import { formatPrettyPercent } from "lib/utils";
+import { divWithDefault, formatPrettyPercent } from "lib/utils";
 
 const StatWithLabel = ({
   label,
@@ -49,15 +49,17 @@ export const ValidatorStats = ({
   singleStakingDenom,
 }: ValidatorStatsProps) => {
   const { data: stakingProvisions, isLoading: isStakingProvisionsLoading } =
-    useValidatorStakingProvisions();
+    useValidatorStakingProvisions(!!singleStakingDenom);
   const { data: delegations, isLoading: isDelegationsLoading } =
     useValidatorDelegators(validatorAddress);
 
   const estimatedApr = stakingProvisions
     ? formatPrettyPercent(
-        stakingProvisions.stakingProvisions
-          .div(totalVotingPower.mul(1 - commissionRate))
-          .toNumber(),
+        divWithDefault(
+          stakingProvisions.stakingProvisions,
+          totalVotingPower.mul(1 - commissionRate),
+          0
+        ).toNumber(),
         2,
         true
       )